Struct cabide::Cabide

pub struct Cabide<T> { /* fields omitted */ }

Abstracts typed database binded to a specific file

Specified type will be (de)serialized from/to the file

If the type changes to have different field order, field types or if more fields are added deserialization may be broken, please keep the type unchanged or migrate the database first

Free blocks in the middle of the file will be cached and prefered, but no data is fragmented over them

use serde::{Serialize, Deserialize};
use cabide::Cabide;

#[derive(Debug, Serialize, Deserialize, PartialEq)]
struct Data {
    name: String,
    maybe_number: Option<u8>,
}

// Opens file pre-filling it, creates it since it's first run
let mut cbd: Cabide<Data> = Cabide::new("test.file", Some(1000))?;
assert_eq!(cbd.blocks()?, 1000);

for _ in 0..100 {
    let data = random_data();
    let primary_key = cbd.write(&data)?;
    assert_eq!(cbd.read(primary_key)?, data);
}

cbd.remove(40)?;
cbd.remove(30)?;
cbd.remove(35)?;

// Since there are empty blocks in the middle of the file we re-use one of them
// (the last one to be available that fits the data)
assert_eq!(cbd.write(&random_data())?, 35);

Implementations

impl<T> Cabide<T>

pub fn new<P>(filename: P, blocks: Option<u64>) -> Result<Self, Error> where
P: AsRef<Path>,

Binds database to specified file, creating it if non existent

Pads file to have specified number of blocks, pre-filling it

If file already exists empty blocks in the middle of it will be cached and prefered, next empty block number will be identified and cached too

use cabide::Cabide;

// Opens file without pre-filling it, creates it since it's first run
let mut cbd: Cabide<u8> = Cabide::new("test2.file", None)?;
assert_eq!(cbd.blocks()?, 0);

// Re-opens file now pre-filling it
cbd = Cabide::new("test2.file", Some(1000))?;
assert_eq!(cbd.blocks()?, 1000);

// Re-opens the file asking for less blocks than available, only to be ignored
cbd = Cabide::new("test2.file", Some(30))?;
assert_eq!(cbd.blocks()?, 1000);

pub fn blocks(&self) -> Result<u64, Error>

Returns number of blocks written to file (some may be empty)

use cabide::Cabide;

// Opens file without pre-filling it, creates it since it's first run
let mut cbd: Cabide<u8> = Cabide::new("test10.file", Some(1000))?;
assert_eq!(cbd.blocks()?, 1000);

pub fn truncate(&mut self) -> Result<(), Error>

impl<T> Cabide<T> where
for<'de> T: Deserialize<'de>,

pub fn remove(&mut self, block: u64) -> Result<T, Error>

Mark object blocks as empty, cacheing them, returns removed content

use cabide::Cabide;

// Opens database, creates it since it's first run
let mut cbd: Cabide<u8> = Cabide::new("test3.file", None)?;

for i in 0..100 {
    cbd.write(&i)?;
}

assert_eq!(cbd.remove(30)?, 30);
assert_eq!(cbd.remove(3)?, 3);

// Writing re-uses freed blocks in the middle of the file (using the last ones first)
assert_eq!(cbd.write(&100)?, 3);
assert_eq!(cbd.remove(3)?, 100);

pub fn read(&mut self, block: u64) -> Result<T, Error>

Returns object deserialized from specified starting block (and its continuations)

use cabide::Cabide;

// Opens file pre-filling it, creates it since it's first run
let mut cbd: Cabide<u8> = Cabide::new("test4.file", None)?;

for i in 0..100 {
    cbd.write(&i)?;
}

assert_eq!(cbd.read(30)?, 30);
assert_eq!(cbd.read(3)?, 3);

pub fn first(&mut self, filter: impl Fn(&T) -> bool) -> Option<T>

Returns first element to be selected by the filter function

Works in O(n), testing each block until the first is found

use serde::{Serialize, Deserialize};
use cabide::Cabide;

#[derive(Debug, Serialize, Deserialize)]
struct Student {
    name: String,
    dre: u64,
    classes: Vec<u16>
}

// Opens file pre-filling it, creates it since it's first run
let mut cbd: Cabide<Student> = Cabide::new("test5.file", Some(1000))?;
assert_eq!(cbd.blocks()?, 1000);

for i in 0..20 {
    cbd.write(&random_student(vec![1023, random(), random()]))?;
}

cbd.write(&Student {
    name: "Mr Legit Student".to_owned(),
    dre: 10101010,
    classes: vec![3],
})?;

let student = cbd.first(|student| student.dre == 10101010).unwrap();
assert_eq!(&student.name, "Mr Legit Student");

pub fn filter(&mut self, filter: impl Fn(&T) -> bool) -> Vec<T>

Returns list of element selected by the filter function

use serde::{Serialize, Deserialize};
use cabide::Cabide;

#[derive(Debug, Serialize, Deserialize)]
struct Student {
    name: String,
    dre: u64,
    classes: Vec<u16>
}

// Opens file pre-filling it, creates it since it's first run
let mut cbd: Cabide<Student> = Cabide::new("test6.file", Some(1000))?;

cbd.write(&Student {
    name: "Mr Legit Student".to_owned(),
    dre: 10101010,
    classes: vec![3],
})?;

for i in 0..20 {
    cbd.write(&random_student(vec![1023, random(), random()]))?;
}

let students = cbd.filter(|student| student.classes.contains(&1023));
assert_eq!(students.len(), 20);

pub fn remove_with(&mut self, filter: impl Fn(&T) -> bool) -> Vec<T>

Sorry, docs are still on their way for this

impl<T: Serialize> Cabide<T>

pub fn write(&mut self, obj: &T) -> Result<u64, Error>

Writes data to database, splitting data in multiple blocks if needed

Re-uses removed blocks, doesn’t fragment data

use cabide::Cabide;

let mut cbd: Cabide<u8> = Cabide::new("test7.file", None)?;

// Each u8 will always fit in one block
for i in 0..1000 {
    assert_eq!(cbd.write(&rand::random())?, i);
}

cbd.remove(30)?;
cbd.remove(58)?;

// Since there are empty blocks in the middle of the file we re-use one of them
// (the last one to be available that fits the data)
assert_eq!(cbd.write(&rand::random())?, 58);